+++ packages/perseus/src/__stories__/server-item-renderer.stories.tsx
@@ -8,7 +8,7 @@ import {
     itemWithLintingError,
     labelImageItem,
     itemWithImages,
-    itemWithMultipleNumericInputs,
+    itemWithMultipleInputNumbers,
     itemWithRadioAndExpressionWidgets,
 } from "../__testdata__/server-item-renderer.testdata";
 import {ServerItemRenderer} from "../server-item-renderer";
@@ -23,7 +23,7 @@ export default {
     title: "Perseus/Renderers/Server Item Renderer",
 } as Story;
 
-export const NumericInputItem = (args: StoryArgs): React.ReactElement => {
+export const InputNumberItem = (args: StoryArgs): React.ReactElement => {
     return <ServerItemRendererWithDebugUI item={itemWithInput} />;
 };
 
@@ -56,7 +56,7 @@ export const InputNumberWithInteractionCallback = (
 ): React.ReactElement => {
     return (
         <ServerItemRendererWithDebugUI
-            item={itemWithMultipleNumericInputs}
+            item={itemWithMultipleInputNumbers}
             apiOptions={{
                 interactionCallback: (data) => {
                     // We are logging the interaction callback data to the console