+++ packages/perseus/src/index.ts
@@ -18,7 +18,6 @@ export {default as ServerItemRenderer} from "./server-item-renderer";
 export {default as HintsRenderer} from "./hints-renderer";
 export {default as HintRenderer} from "./hint-renderer";
 export {default as Renderer} from "./renderer";
-export {scorePerseusItem} from "./renderer-util";
 
 /**
  * Widgets
@@ -101,6 +100,7 @@ export {
     containerSizeClass,
     getInteractiveBoxFromSizeClass,
 } from "./util/sizing-utils";
+export {mathOnlyParser} from "./widgets/interactive-graphs/utils";
 export {
     getAnswersFromWidgets,
     injectWidgets,
@@ -208,9 +208,8 @@ export type {
     Size,
     CollinearTuple,
     MathFormat,
+    InputNumberWidget, // TODO(jeremy): remove?
     PerseusArticle,
-    InputNumberWidget, // Used for usurpation of InputNumberWidget in perseus-editor
-    NumericInputWidget, // Used for usurpation of InputNumberWidget in perseus-editor
     // Widget configuration types
     PerseusImageBackground,
     PerseusInputNumberWidgetOptions,
@@ -227,7 +226,6 @@ export type {
     PerseusWidgetsMap,
     MultiItem,
 } from "./perseus-types";
-export type {UserInputMap} from "./validation.types";
 export type {Coord} from "./interactive2/types";
 export type {MarkerType} from "./widgets/label-image/types";