+++ packages/perseus/src/util/parse-perseus-json/perseus-parsers/matrix-widget.ts
@@ -4,8 +4,10 @@ import {
     constant,
     number,
     object,
+    optional,
     string,
 } from "../general-purpose-parsers";
+import {defaulted} from "../general-purpose-parsers/defaulted";
 
 import {parseWidget} from "./widget";
 
@@ -13,13 +15,13 @@ import type {MatrixWidget} from "../../../perseus-types";
 import type {Parser} from "../parser-types";
 
 export const parseMatrixWidget: Parser<MatrixWidget> = parseWidget(
-    constant("matrix"),
+    defaulted(constant("matrix"), () => "matrix"),
     object({
-        prefix: string,
-        suffix: string,
+        prefix: optional(string),
+        suffix: optional(string),
         answers: array(array(number)),
-        cursorPosition: array(number),
+        cursorPosition: optional(array(number)),
         matrixBoardSize: array(number),
-        static: boolean,
+        static: optional(boolean),
     }),
 );