+++ packages/perseus/src/widgets/interactive-graphs/interactive-graph.testdata.ts
@@ -162,6 +162,25 @@ export const polygonQuestion: PerseusRenderer =
         })
         .build();
 
+export const unlimitedPolygonQuestion: PerseusRenderer =
+    interactiveGraphQuestionBuilder()
+        .withContent(
+            "**Sides shown** Drag the vertices of the triangle below to draw a right triangle with side lengths $3$, $4$, and $5$. \n[[\u2603 interactive-graph 1]] \n",
+        )
+        .withGridStep(1, 1)
+        .withSnapStep(0.25, 0.25)
+        .withTickStep(1, 1)
+        .withXRange(-10, 10)
+        .withYRange(-10, 10)
+        .withPolygon("grid", {
+            match: "congruent",
+            numSides: "unlimited",
+            showSides: true,
+            showAngles: true,
+            coords: [],
+        })
+        .build();
+
 export const polygonWithStartingCoordsQuestion: PerseusRenderer =
     interactiveGraphQuestionBuilder()
         .withPolygon("grid", {
@@ -284,6 +303,8 @@ export const polygonWithUnlimitedSidesQuestion: PerseusRenderer =
             "**Example of unlimited polygon sides** \n[[\u2603 interactive-graph 1]] \n",
         )
         .withPolygon("grid", {
+            showAngles: true,
+            showSides: true,
             numSides: "unlimited",
             coords: [
                 [0, 0],